stress-sleep: ensure residency_ns is computed using unsigned long multiplication
Currently the multiply is using int types, cast to unsigned long to
ensure no overflow can occur.

+++ stress-sleep.c
@@ -113,7 +113,7 @@ static void *stress_pthread_func(void *c)
 		t1 = stress_time_now();
 		for (cc = cstate_list; cc; cc = cc->next) {
 			if (cc->residency > 0) {
-				const unsigned long residency_ns = (cc->residency + 1) * 1000;
+				const unsigned long residency_ns = (unsigned long)(cc->residency + 1) * 1000UL;
 
 				expected += cc->residency;
 				tv.tv_nsec = residency_ns % 1000000000;
